Login - OK Redirect to Dashboard - OK Create Project - OK Logout - OK

+++ database-config.js
@@ -53,6 +53,8 @@ function databaseConnected() {
 
     console.log(messages.databaseConnected);
 
+    if(true) return; // Setar para falso para recriar as tabelas. TODO Remover e colocar algo mais elegante
+
     User.sync({ force: true }).then(() => {console.log(messages.usersTableCreated)});
     Project.sync({ force: true }).then(() => {console.log(messages.projectsTableCreated)});
